Fix lint and test
